# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Restaurant Documentation",
        "description": "L5 Swagger OpenApi description for restaurant project",
        "contact": {
            "email": "drrifatalashwad0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/v1.0/customers": {
            "get": {
                "tags": [
                    "reports"
                ],
                "summary": "This method is to get  bookings ",
                "description": "This method is to get bookings",
                "operationId": "getCustomers",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": true,
                        "example": ""
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter by user creation start date (YYYY-MM-DD)",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter by user creation end date (YYYY-MM-DD)",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "search_key",
                        "in": "query",
                        "description": "Keyword for searching users by name, email, or phone",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter by review rating",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "frequency_visit",
                        "in": "query",
                        "description": "Visit frequency category (New, Regular, VIP)",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "review_start_date",
                        "in": "query",
                        "description": "Filter reviews by start date (YYYY-MM-DD)",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "review_end_date",
                        "in": "query",
                        "description": "Filter reviews by end date (YYYY-MM-DD)",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "review_keyword",
                        "in": "query",
                        "description": "Keyword to search within review comments",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "frequency_visit",
                        "in": "query",
                        "description": "Filter users based on visit frequency",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Comma-separated list of booking statuses",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "description": "Comma-separated list of payment statuses",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "sub_service_ids",
                        "in": "query",
                        "description": "Comma-separated list of sub-service IDs",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "duration_in_minute",
                        "in": "query",
                        "description": "Filter by booking duration in minutes",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "booking_type",
                        "in": "query",
                        "description": "Comma-separated list of booking types",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "date_filter",
                        "in": "query",
                        "description": "Filter bookings by date (e.g., today, this_week, previous_week, this_month, etc.)",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter users by name",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filter users by email",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Filter users by phone number",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "last_visited_date",
                        "in": "query",
                        "description": "Filter users by last visited date (YYYY-MM-DD)",
                        "required": false,
                        "example": ""
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort order for users (ASC or DESC)",
                        "required": false,
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocesseble Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}